Splited task-summary-panel and task-summary-accordion
